+++ Task-10_Browser-game/src/js/controllers/components/menuComponent.js
@@ -1,6 +1,6 @@
-import { HTMLTags } from '../render.js'
-import { Commands } from '../controls.js'
-import { IndexEnumerator } from '../utils.js'
+import { HTMLTags } from '../../render.js'
+import { Commands } from '../../controls.js'
+import { IndexEnumerator } from '../../utils.js'
 
 const defaultstyleClasses = {
     table: 'width_100 align_center',
